+++ CHANGELOG.md
@@ -11,7 +11,7 @@
     sudo npm i vue-toastification@next
     ```
     - add event emmited on form submit on add transaction
-    - done adding transaction + success toast
+    - done add/remove transaction + success toast
 
 - 2024-03-07
     - init